Removed warnings. Improved CMake files.

+++ VKTS/src/layer0/binary_buffer/BinaryBuffer.cpp
@@ -133,7 +133,7 @@ VkBool32 BinaryBuffer::seek(const int64_t offset, const VkTsSearch search)
                 return VK_FALSE;
             }
 
-            pos = offset;
+            pos = (size_t)offset;
 
             return VK_TRUE;
         }
@@ -155,7 +155,7 @@ VkBool32 BinaryBuffer::seek(const int64_t offset, const VkTsSearch search)
                 }
             }
 
-            pos += offset;
+            pos += (size_t)offset;
 
             return VK_TRUE;
         }
